Finished MLA tests. - Added dynamic random tests. - Improved output.
